account/requirements.txt
grpcio-tools ~= 1.40
psycopg2-binary ~=2.9